Cast ceil return to integer
The manual says:

> The return value of ceil() is still of type float as the value range
 of float is usually bigger than that of integer.

+++ src/Datagrid/Pager.php
@@ -76,7 +76,7 @@ class Pager extends BasePager
         } else {
             $offset = ($this->getPage() - 1) * $this->getMaxPerPage();
 
-            $this->setLastPage(ceil($this->getNbResults() / $this->getMaxPerPage()));
+            $this->setLastPage((int) ceil($this->getNbResults() / $this->getMaxPerPage()));
 
             $this->getQuery()->setFirstResult($offset);
             $this->getQuery()->setMaxResults($this->getMaxPerPage());
